Evolution pdf seperated
Evolution pdfs are now its seperate thing. So none of those gets scrapped and updated in the update_qdrant_since function. It now has its own function to lower money cost and update time, since these pdfs were big and doesnt need to be updated everything u change a little thing on one of the pages its on the web.